+++ pages/calendar/index.tsx
@@ -1,11 +1,11 @@
 import Calendar from '@/components/calendar-compound/calendar';
 import styles from './calendar.module.scss';
 import CalendarContainer from '@/components/calendar-monthly/calendar-container';
-import CalendarTodo from '@/components/calendar-monthly/calendar-todo';
 import CreateTodoButton from '@/components/calendar-monthly/create-todo-button';
 import CalendarTodoDate from '@/components/calendar-monthly/calendar-todo-date';
 import CalendarInstance from '@/utils/date/date.utils';
 import useSelect from '@/hooks/use-select';
+import CalendarTodos from '@/components/calendar-monthly/calendar-todos';
 
 export type CalendarProps = {
   year: number | null;
@@ -51,7 +51,7 @@ export default function CalendarPage() {
         </div>
         <CalendarContainer />
         <CalendarTodoDate />
-        <CalendarTodo />
+        <CalendarTodos />
         <CreateTodoButton />
       </Calendar>
     </main>
